requirements.txt
Flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
geopy==2.4.1 